Fix spec regressions; Mostly clear DB
There were a ton of spec errors, parser returning the wrong thing. I assumed I just screwed
everything up while refactoring but I guess it was a problem with the db not getting cleared or...
I dunno, when I added a db.clear to the spec helper, all the other problems went away.